requirements.txt
pyserial
matplotlib
pynmea2
pyqt5
seaborn
vincenty